+++ example/fork.js
@@ -1,0 +1,4 @@
+const [,, ...args] = process.argv
+console.log(args)
+console.error(process.env.EXAMPLE)
+process.exit(5)