+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 16 February 2019
+
+### 1.1.0
+
+- [doc] Document example.
+- [feature] Pass the `input` and other properties in the context.
+
 ## 24 October 2018
 
 ### 1.0.0